moved player constants
