+++ TentakelsAttacking2/Constants/public/CPlayer.hpp
@@ -1,0 +1,11 @@
+//
+// Purpur Tentakel
+// 27.10.2022
+//
+
+#pragma once
+
+struct CPlayer {
+	size_t minPlayerCount = 2;
+	size_t maxPlayerCount = 7;
+};
