+++ TentakelsAttacking2/TentakelsAttacking2.vcxproj.filters
@@ -189,7 +189,7 @@
     <ClInclude Include="UI\Elements\Default\public\Line.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Constants\public\CTextProcecingConstant.hpp">
+    <ClInclude Include="Constants\public\CTextProcecing.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="UI\public\PopUpManager.h">
@@ -232,6 +232,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Constants\public\CConstants.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Constants\public\CPlayer.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
